Fixes and improvements to flows
