+++ src/github/dashboardWebviewProvider.ts
@@ -7,11 +7,11 @@ import * as vscode from 'vscode';
 import Logger from '../common/logger';
 import { ITelemetry } from '../common/telemetry';
 import { getNonce, IRequestMessage, WebviewBase } from '../common/webview';
-import { ChatSessionWithPR } from './copilotApi';
 import { CopilotRemoteAgentManager } from './copilotRemoteAgent';
 import { FolderRepositoryManager, ReposManagerState } from './folderRepositoryManager';
 import { IssueModel } from './issueModel';
 import { RepositoriesManager } from './repositoriesManager';
+import { SessionData, TaskManager } from './taskManager';
 
 // Dashboard state discriminated union
 export type DashboardState = DashboardLoading | DashboardReady;
@@ -35,19 +35,6 @@ export interface DashboardData {
 	issueQuery: string;
 }
 
-export interface SessionData {
-	id: string;
-	title: string;
-	status: string;
-	dateCreated: string;
-	isCurrentBranch?: boolean;
-	pullRequest?: {
-		number: number;
-		title: string;
-		url: string;
-	};
-}
-
 export interface IssueData {
 	number: number;
 	title: string;
@@ -67,6 +54,7 @@ export class DashboardWebviewProvider extends WebviewBase {
 
 	private _issueQuery: string;
 	private _repos?: string[];
+	private _taskManager: TaskManager;
 
 	constructor(
 		private readonly _context: vscode.ExtensionContext,
@@ -83,6 +71,7 @@ export class DashboardWebviewProvider extends WebviewBase {
 		this._webview = panel.webview;
 		this._issueQuery = issueQuery || 'is:open assignee:@me milestone:"September 2025"';
 		this._repos = repos;
+		this._taskManager = new TaskManager(this._repositoriesManager, this._copilotRemoteAgentManager);
 		super.initialize();
 
 		// Set webview options
@@ -195,63 +184,16 @@ export class DashboardWebviewProvider extends WebviewBase {
 	}
 
 	private async getActiveSessions(): Promise<SessionData[]> {
-		try {
-			// Create a cancellation token for the request
-			const source = new vscode.CancellationTokenSource();
-			const token = source.token;
-
-			const sessions = await this._copilotRemoteAgentManager.provideChatSessions(token);
-			let filteredSessions = sessions;
-
-			// Filter sessions by repositories if specified
-			const targetRepos = this.getTargetRepositories();
-			if (targetRepos.length > 0) {
-				filteredSessions = sessions.filter(session => {
-					// If session has a pull request, check if it belongs to one of the target repos
-					if (session.pullRequest?.html_url) {
-						const urlMatch = session.pullRequest.html_url.match(/github\.com\/([^\/]+)\/([^\/]+)\/pull\/\d+/);
-						if (urlMatch) {
-							const [, owner, repo] = urlMatch;
-							const repoIdentifier = `${owner}/${repo}`;
-							return targetRepos.some(targetRepo =>
-								targetRepo.toLowerCase() === repoIdentifier.toLowerCase()
-							);
-						}
-					}
-					// If no pull request or URL doesn't match pattern, include it
-					// (this covers sessions that might not be tied to a specific repo)
-					return targetRepos.length === 0;
-				});
-			}
-
-			// Deduplicate sessions by ID to prevent duplicate display
-			const sessionMap = new Map<string, SessionData>();
-			for (const session of filteredSessions) {
-				const sessionData = this.convertSessionToData(session);
-				sessionMap.set(sessionData.id, sessionData);
-			}
-
-			return Array.from(sessionMap.values());
-		} catch (error) {
-			Logger.error(`Failed to get active sessions: ${error}`, DashboardWebviewProvider.ID);
-			return [];
-		}
+		const targetRepos = this.getTargetRepositories();
+		return await this._taskManager.getActiveSessions(targetRepos);
 	}
 
-	private convertSessionToData(session: ChatSessionWithPR): SessionData {
-		const isCurrentBranch = this.isSessionAssociatedWithCurrentBranch(session);
-		return {
-			id: session.id,
-			title: session.label,
-			status: session.status ? session.status.toString() : 'Unknown',
-			dateCreated: session.timing?.startTime ? new Date(session.timing.startTime).toISOString() : '',
-			isCurrentBranch,
-			pullRequest: session.pullRequest ? {
-				number: session.pullRequest.number,
-				title: session.pullRequest.title,
-				url: session.pullRequest.html_url
-			} : undefined
-		};
+	private async switchToLocalTask(branchName: string): Promise<void> {
+		await this._taskManager.switchToLocalTask(branchName);
+		// Update dashboard to reflect current branch change
+		setTimeout(() => {
+			this.updateDashboard();
+		}, 500);
 	}
 
 	private async getMilestoneIssues(): Promise<IssueData[]> {
@@ -405,10 +347,32 @@ export class DashboardWebviewProvider extends WebviewBase {
 			case 'open-pull-request':
 				await this.openPullRequest(message.args?.pullRequest);
 				break;
+			case 'switch-to-local-task':
+				await this.switchToLocalTask(message.args?.branchName);
+				break;
 			default:
 				await super._onDidReceiveMessage(message);
 				break;
 		}
+	}
+
+	/**
+	 * Creates a temporary session that shows in the dashboard with a loading state
+	 */
+	private createTemporarySession(query: string, type: 'local' | 'remote'): string {
+		const tempId = this._taskManager.createTemporarySession(query, type);
+		// Immediately update the dashboard to show the temporary session
+		this.updateDashboard();
+		return tempId;
+	}	/**
+	 * Removes a temporary session from the dashboard
+	 */
+	private removeTemporarySession(tempId: string): void {
+		this._taskManager.removeTemporarySession(tempId);
+		// Update dashboard to remove the temporary session
+		setTimeout(() => {
+			this.updateDashboard();
+		}, 500); // Small delay to allow real session to be created
 	}
 
 	private async handleChatSubmission(query: string): Promise<void> {
@@ -441,16 +405,32 @@ export class DashboardWebviewProvider extends WebviewBase {
 
 			// Check if user explicitly mentions @copilot for remote background session
 			if (query.includes('@copilot')) {
+				// Create temporary session first
+				const tempId = this.createTemporarySession(enhancedQuery.replace(/@copilot\s*/, '').trim(), 'remote');
+
 				// Create background session using chat sessions API
-				await this.createRemoteBackgroundSession(enhancedQuery);
+				try {
+					await this.createRemoteBackgroundSession(enhancedQuery);
+				} finally {
+					// Remove temporary session regardless of success/failure
+					this.removeTemporarySession(tempId);
+				}
 				return;
 			}
 
 			// Check if user explicitly mentions @local for local workflow
 			if (query.includes('@local')) {
-				// Remove @local prefix and set up local workflow directly
+				// Create temporary session first
 				const cleanQuery = enhancedQuery.replace(/@local\s*/, '').trim();
-				await this.setupLocalWorkflow(cleanQuery || enhancedQuery);
+				const tempId = this.createTemporarySession(cleanQuery || enhancedQuery, 'local');
+
+				// Remove @local prefix and set up local workflow directly
+				try {
+					await this.setupLocalWorkflow(cleanQuery || enhancedQuery);
+				} finally {
+					// Remove temporary session regardless of success/failure
+					this.removeTemporarySession(tempId);
+				}
 				return;
 			}
 
@@ -462,11 +442,27 @@ export class DashboardWebviewProvider extends WebviewBase {
 				const workMode = await this.showWorkModeQuickPick();
 
 				if (workMode === 'remote') {
+					// Create temporary session first
+					const tempId = this.createTemporarySession(enhancedQuery, 'remote');
+
 					// Use @copilot to start a new chat session
-					await vscode.commands.executeCommand('workbench.action.chat.open', { query: `@copilot ${enhancedQuery}` });
+					try {
+						await vscode.commands.executeCommand('workbench.action.chat.open', { query: `@copilot ${enhancedQuery}` });
+					} finally {
+						// Remove temporary session
+						this.removeTemporarySession(tempId);
+					}
 				} else if (workMode === 'local') {
+					// Create temporary session first
+					const tempId = this.createTemporarySession(enhancedQuery, 'local');
+
 					// Create a new branch and set up local chat with agent mode
-					await this.setupLocalWorkflow(enhancedQuery);
+					try {
+						await this.setupLocalWorkflow(enhancedQuery);
+					} finally {
+						// Remove temporary session
+						this.removeTemporarySession(tempId);
+					}
 				}
 				// If workMode is undefined, user cancelled - do nothing
 			} else {
@@ -828,395 +824,35 @@ export class DashboardWebviewProvider extends WebviewBase {
 	 * Determines if a query represents a coding task vs a general question using VS Code's Language Model API
 	 */
 	private async isCodingTask(query: string): Promise<boolean> {
-		try {
-			// Try to get a language model for classification
-			const models = await vscode.lm.selectChatModels({
-				vendor: 'copilot',
-				family: 'gpt-4o'
-			});
-
-			if (!models || models.length === 0) {
-				// Fallback to keyword-based classification if no LM available
-				return this.isCodingTaskFallback(query);
-			}
-
-			const model = models[0];
-
-			// Create a focused prompt for binary classification
-			const classificationPrompt = `You are a classifier that determines whether a user query represents a coding/development task or a general question.
-
-Examples of CODING TASKS:
-- "Implement user authentication"
-- "Fix the bug in the login function"
-- "Add a search feature to the app"
-- "Refactor the database connection code"
-- "Create unit tests for the API"
-- "Debug the memory leak issue"
-- "Update the CSS styling"
-- "Build a REST endpoint"
-
-Examples of GENERAL QUESTIONS:
-- "How does authentication work?"
-- "What is a REST API?"
-- "Explain the difference between async and sync"
-- "What are the benefits of unit testing?"
-- "How do I learn React?"
-- "What is the best IDE for Python?"
-
-Respond with exactly one word: "CODING" if the query is about implementing, building, fixing, creating, or working on code. "GENERAL" if it's asking for information, explanations, or learning resources.
-
-Query: "${query}"
-
-Classification:`;
-
-			const messages = [vscode.LanguageModelChatMessage.User(classificationPrompt)];
-
-			const response = await model.sendRequest(messages, {
-				justification: 'Classifying user query type for workflow routing'
-			});
-
-			let result = '';
-			for await (const chunk of response.text) {
-				result += chunk;
-			}
-
-			// Parse the response - look for "CODING" or "GENERAL"
-			const cleanResult = result.trim().toUpperCase();
-			return cleanResult.includes('CODING');
-
-		} catch (error) {
-			Logger.error(`Failed to classify query using LM API: ${error}`, DashboardWebviewProvider.ID);
-			// Fallback to keyword-based classification
-			return this.isCodingTaskFallback(query);
-		}
+		return await this._taskManager.isCodingTask(query);
 	}
 
 	/**
 	 * Fallback keyword-based classification when LM API is unavailable
 	 */
-	private isCodingTaskFallback(query: string): boolean {
-		const codingKeywords = [
-			'implement', 'create', 'add', 'build', 'develop', 'code', 'write',
-			'fix', 'debug', 'resolve', 'solve', 'repair',
-			'refactor', 'optimize', 'improve', 'enhance', 'update',
-			'feature', 'function', 'method', 'class', 'component',
-			'api', 'endpoint', 'service', 'module', 'library',
-			'test', 'testing', 'unit test', 'integration test',
-			'bug', 'issue', 'error', 'exception', 'crash'
-		];
-
-		const lowercaseQuery = query.toLowerCase();
-		return codingKeywords.some(keyword => lowercaseQuery.includes(keyword));
-	}
-
 	/**
 	 * Shows a quick pick to let user choose between local and remote work
 	 */
 	private async showWorkModeQuickPick(): Promise<'local' | 'remote' | undefined> {
-		const quickPick = vscode.window.createQuickPick();
-		quickPick.title = 'Choose how to work on this task';
-		quickPick.placeholder = 'Select whether to work locally or remotely';
-		quickPick.items = [
-			{
-				label: '$(device-desktop) Work locally',
-				detail: 'Create a new branch and work in your local environment',
-				alwaysShow: true
-			},
-			{
-				label: '$(cloud) Work remotely',
-				detail: 'Use GitHub Copilot remote agent to work in the cloud',
-				alwaysShow: true
-			}
-		];
-
-		return new Promise<'local' | 'remote' | undefined>((resolve) => {
-			quickPick.onDidAccept(() => {
-				const selectedItem = quickPick.selectedItems[0];
-				quickPick.hide();
-				if (selectedItem) {
-					if (selectedItem.label.includes('locally')) {
-						resolve('local');
-					} else if (selectedItem.label.includes('remotely')) {
-						resolve('remote');
-					}
-				}
-				resolve(undefined);
-			});
-
-			quickPick.onDidHide(() => {
-				quickPick.dispose();
-				resolve(undefined);
-			});
-
-			quickPick.show();
-		});
+		return await this._taskManager.showWorkModeQuickPick();
 	}
 
 	/**
 	 * Sets up local workflow: creates branch and opens chat with agent mode
 	 */
 	private async setupLocalWorkflow(query: string): Promise<void> {
-		try {
-			await vscode.window.withProgress({
-				location: vscode.ProgressLocation.Notification,
-				title: 'Setting up local workflow',
-				cancellable: false
-			}, async (progress) => {
-				progress.report({ message: 'Initializing...', increment: 10 });
-
-				// Get the first available folder manager with a repository
-				const folderManager = this._repositoriesManager.folderManagers.find(fm =>
-					fm.gitHubRepositories.length > 0
-				);
-
-				if (!folderManager) {
-					vscode.window.showErrorMessage('No GitHub repository found in the current workspace.');
-					return;
-				}
-
-				// Generate a logical branch name from the query
-				progress.report({ message: 'Generating branch name...', increment: 30 });
-				const branchName = await this.generateBranchName(query);
-
-				// Create the branch
-				progress.report({ message: `Creating branch: ${branchName}`, increment: 60 });
-				await folderManager.repository.createBranch(branchName, true);
-
-				// Create a fresh chat session and open with ask mode
-				progress.report({ message: 'Opening chat session...', increment: 90 });
-				await vscode.commands.executeCommand('workbench.action.chat.newChat');
-				await vscode.commands.executeCommand('workbench.action.chat.open', {
-					query,
-					mode: 'ask'
-				});
-
-				progress.report({ message: 'Local workflow ready!', increment: 100 });
-			});
-
-			// Show success message
-			vscode.window.showInformationMessage('Local workflow setup complete! Ready to work on your task.');
-
-		} catch (error) {
-			Logger.error(`Failed to setup local workflow: ${error}`, DashboardWebviewProvider.ID);
-			vscode.window.showErrorMessage(`Failed to setup local workflow: ${error}`);
-		}
+		await this._taskManager.setupLocalWorkflow(query);
 	}
 
 	/**
 	 * Creates a remote background session using the copilot remote agent
 	 */
 	private async createRemoteBackgroundSession(query: string): Promise<void> {
-		try {
-			await vscode.window.withProgress({
-				location: vscode.ProgressLocation.Notification,
-				title: 'Creating remote coding agent session',
-				cancellable: false
-			}, async (progress) => {
-				progress.report({ message: 'Preparing task...', increment: 20 });
-
-				// Extract the query without @copilot mention
-				const cleanQuery = query.replace(/@copilot\s*/, '').trim();
-
-				// Use the copilot remote agent manager to create a new session
-				progress.report({ message: 'Creating session with coding agent...', increment: 60 });
-				const sessionResult = await this._copilotRemoteAgentManager.provideNewChatSessionItem({
-					request: {
-						prompt: cleanQuery,
-						references: [],
-						participant: 'copilot-swe-agent'
-					} as any, // Type assertion as we're using this internal API
-					prompt: cleanQuery,
-					history: [],
-					metadata: { source: 'dashboard' }
-				}, new vscode.CancellationTokenSource().token);
-
-				// Show confirmation that the task has been created
-				if (sessionResult && sessionResult.id) {
-					progress.report({ message: 'Session created successfully!', increment: 90 });
-
-					const sessionTitle = sessionResult.label || `Session ${sessionResult.id}`;
-					const viewAction = 'View Session';
-					const result = await vscode.window.showInformationMessage(
-						`Created new coding agent task: ${sessionTitle}`,
-						viewAction
-					);
-
-					if (result === viewAction) {
-						// Open the session if user chooses to view it
-						await vscode.window.showChatSession('copilot-swe-agent', sessionResult.id, {});
-					}
-
-					// Refresh the dashboard to show the new session
-					setTimeout(() => {
-						this.updateDashboard();
-					}, 1000);
-
-					progress.report({ message: 'Remote session ready!', increment: 100 });
-				} else {
-					vscode.window.showErrorMessage('Failed to create coding agent session.');
-				}
-			});
-
-		} catch (error) {
-			Logger.error(`Failed to create remote background session: ${error}`, DashboardWebviewProvider.ID);
-			vscode.window.showErrorMessage(`Failed to create coding agent session: ${error}`);
-		}
-	}
-
-	/**
-	 * Generates a logical branch name from a query using LM API with uniqueness checking
-	 */
-	private async generateBranchName(query: string): Promise<string> {
-		try {
-			// Try to get a language model for branch name generation
-			const models = await vscode.lm.selectChatModels({
-				vendor: 'copilot',
-				family: 'gpt-4o'
-			});
-
-			let baseName: string;
-
-			if (models && models.length > 0) {
-				const model = models[0];
-
-				// Create a focused prompt for branch name generation
-				const namePrompt = `Generate a concise, descriptive git branch name for this task. The name should be:
-- 3-6 words maximum
-- Use kebab-case (lowercase with hyphens)
-- Be descriptive but brief
-- Follow conventional branch naming patterns
-- No special characters except hyphens
-- Start with a category like "feature/", "fix/", "refactor/", etc. if appropriate
-
-Examples:
-- "implement user authentication" → "feature/user-authentication"
-- "fix login bug" → "fix/login-bug"
-- "add search functionality" → "feature/search-functionality"
-- "refactor database code" → "refactor/database-code"
-- "update styling" → "update/styling"
-
-Task: "${query}"
-
-Branch name:`;
-
-				const messages = [vscode.LanguageModelChatMessage.User(namePrompt)];
-
-				const response = await model.sendRequest(messages, {
-					justification: 'Generating descriptive branch name for development task'
-				});
-
-				let result = '';
-				for await (const chunk of response.text) {
-					result += chunk;
-				}
-
-				// Clean up the LM response
-				baseName = result.trim()
-					.replace(/^["']|["']$/g, '') // Remove quotes
-					.replace(/[^\w\s/-]/g, '') // Remove special characters except hyphens and slashes
-					.replace(/\s+/g, '-') // Replace spaces with hyphens
-					.replace(/-+/g, '-') // Replace multiple hyphens with single hyphen
-					.replace(/^-|-$/g, '') // Remove leading/trailing hyphens
-					.toLowerCase();
-
-				// Ensure it has a reasonable length
-				if (baseName.length > 50) {
-					baseName = baseName.substring(0, 50).replace(/-[^-]*$/, ''); // Cut at word boundary
-				}
-			} else {
-				// Fallback to simple name generation if LM is unavailable
-				baseName = this.generateFallbackBranchName(query);
-			}
-
-			// Ensure uniqueness by checking existing branches
-			return await this.ensureUniqueBranchName(baseName);
-
-		} catch (error) {
-			Logger.error(`Failed to generate branch name using LM: ${error}`, DashboardWebviewProvider.ID);
-			// Fallback to simple name generation
-			const fallbackName = this.generateFallbackBranchName(query);
-			return await this.ensureUniqueBranchName(fallbackName);
-		}
-	}
-
-	/**
-	 * Fallback branch name generation when LM API is unavailable
-	 */
-	private generateFallbackBranchName(query: string): string {
-		// Clean up the query to create a branch-friendly name
-		const cleaned = query
-			.toLowerCase()
-			.replace(/[^\w\s-]/g, '') // Remove special characters except hyphens
-			.replace(/\s+/g, '-') // Replace spaces with hyphens
-			.replace(/-+/g, '-') // Replace multiple hyphens with single hyphen
-			.replace(/^-|-$/g, ''); // Remove leading/trailing hyphens
-
-		// Truncate to reasonable length
-		const truncated = cleaned.length > 40 ? cleaned.substring(0, 40) : cleaned;
-
-		return `task/${truncated}`;
-	}
-
-	/**
-	 * Ensures branch name uniqueness by checking existing branches and adding suffix if needed
-	 */
-	private async ensureUniqueBranchName(baseName: string): Promise<string> {
-		try {
-			// Get the first available folder manager with a repository
-			const folderManager = this._repositoriesManager.folderManagers.find(fm =>
-				fm.gitHubRepositories.length > 0
-			);
-
-			if (!folderManager) {
-				// If no repository available, just add timestamp for uniqueness
-				const timestamp = Date.now().toString().slice(-6);
-				return `${baseName}-${timestamp}`;
-			}
-
-			// Get existing branch names
-			let existingBranches = new Set<string>();
-			try {
-				if (folderManager.repository.getRefs) {
-					const refs = await folderManager.repository.getRefs({
-						contains: undefined,
-						count: undefined,
-						pattern: undefined,
-						sort: undefined
-					});
-					existingBranches = new Set(
-						refs
-							.filter(ref => ref.type === 1 && ref.name) // RefType.Head and has name
-							.map(ref => ref.name!)
-					);
-				}
-			} catch (error) {
-				Logger.debug(`Could not fetch branch refs: ${error}`, DashboardWebviewProvider.ID);
-				// Continue with empty set - will use timestamp for uniqueness
-			}
-
-			// Check if base name is unique
-			if (!existingBranches.has(baseName)) {
-				return baseName;
-			}
-
-			// If not unique, try adding numeric suffixes
-			for (let i = 2; i <= 99; i++) {
-				const candidateName = `${baseName}-${i}`;
-				if (!existingBranches.has(candidateName)) {
-					return candidateName;
-				}
-			}
-
-			// If still not unique after 99 attempts, add timestamp
-			const timestamp = Date.now().toString().slice(-6);
-			return `${baseName}-${timestamp}`;
-
-		} catch (error) {
-			Logger.error(`Failed to check branch uniqueness: ${error}`, DashboardWebviewProvider.ID);
-			// Fallback to timestamp-based uniqueness
-			const timestamp = Date.now().toString().slice(-6);
-			return `${baseName}-${timestamp}`;
-		}
+		await this._taskManager.createRemoteBackgroundSession(query);
+		// Refresh the dashboard to show the new session
+		setTimeout(() => {
+			this.updateDashboard();
+		}, 1000);
 	}
 
 	private registerBranchChangeListeners(): void {
@@ -1276,55 +912,6 @@ Branch name:`;
 	}
 
 	private _branchChangeTimeout: NodeJS.Timeout | undefined;
-
-	private isSessionAssociatedWithCurrentBranch(session: ChatSessionWithPR): boolean {
-		if (!session.pullRequest) {
-			return false;
-		}
-
-		// Parse the PR URL to get owner and repo
-		const urlMatch = session.pullRequest.html_url.match(/github\.com\/([^\/]+)\/([^\/]+)\/pull\/(\d+)/);
-		if (!urlMatch) {
-			return false;
-		}
-
-		const [, owner, repo] = urlMatch;
-		const prNumber = session.pullRequest.number;
-
-		// Check if any folder manager has this PR checked out on current branch
-		for (const folderManager of this._repositoriesManager.folderManagers) {
-			// Check if this folder manager manages the target repo
-			if (this.folderManagerMatchesRepo(folderManager, owner, repo)) {
-				// Check if the current branch corresponds to this PR
-				const currentBranchName = folderManager.repository.state.HEAD?.name;
-				if (currentBranchName) {
-					// Try to find the PR model for this session
-					try {
-						// Use the active PR if it matches
-						if (folderManager.activePullRequest?.number === prNumber) {
-							return true;
-						}
-						// Also check if the branch name suggests it's a PR branch
-						// Common patterns: pr-123, pull/123, etc.
-						const prBranchPatterns = [
-							new RegExp(`pr-${prNumber}$`, 'i'),
-							new RegExp(`pull/${prNumber}$`, 'i'),
-							new RegExp(`pr/${prNumber}$`, 'i')
-						];
-						for (const pattern of prBranchPatterns) {
-							if (pattern.test(currentBranchName)) {
-								return true;
-							}
-						}
-					} catch (error) {
-						// Ignore errors in checking PR association
-					}
-				}
-			}
-		}
-
-		return false;
-	}
 
 	private getHtmlForWebview(): string {
 		const nonce = getNonce();
